+++ tests/testthat/test_C_sub_dann.R
@@ -1,8 +1,6 @@
-context("testing sub_dann")
-library(dann)
-library(mlbench)
-library(magrittr)
-library(dplyr)
+library(mlbench, warn.conflicts = FALSE)
+library(magrittr, warn.conflicts = FALSE)
+library(dplyr, warn.conflicts = FALSE)
 
 ###############################################
 # Easy problems
@@ -84,7 +82,7 @@ yTest <- test %>%
   as.numeric() %>%
   as.vector()
 
-subDannPreds <- sub_dann(xTrain, yTrain, xTest, 1, 50, 1, FALSE, FALSE, "mcd", 3)
+subDannPreds <- sub_dann(xTrain, yTrain, xTest, 5, 60, 1, FALSE, FALSE, "mcd", 3)
 
 test_that("Validate structure", {
   expect_true(is.vector(subDannPreds))
@@ -160,28 +158,6 @@ test_that("Confirm class probabilities are divisible by k", {
 # K equal 2
 ######################
 K <- 2
-subDannPreds <- sub_dann(xTrain, yTrain, xTest, K, 15, 1, TRUE, FALSE, "mcd", 2)
-test_that("Validate structure", {
-  expect_true(is.matrix(subDannPreds))
-  expect_true(is.numeric(subDannPreds))
-  expect_true(nrow(subDannPreds) == nrow(xTest))
-  expect_true(ncol(subDannPreds) == 2)
-  expect_true(all(colnames(subDannPreds) == c("Class1", "Class2")))
-})
-
-test_that("Confirm class probabilities sum to 1", {
-  expect_true(all(rowSums(subDannPreds) == 1))
-})
-
-possibleProb <- 0:K / K
-test_that("Confirm class probabilities are divisible by k", {
-  expect_true(all(subDannPreds %in% possibleProb))
-})
-
-######################
-# K equal 3
-######################
-K <- 3
 subDannPreds <- sub_dann(xTrain, yTrain, xTest, K, 15, 1, TRUE, FALSE, "mcd", 2)
 test_that("Validate structure", {
   expect_true(is.matrix(subDannPreds))
@@ -310,7 +286,6 @@ test_that("Validate structure", {
   expect_true(is.vector(subDannPreds))
   expect_true(is.numeric(subDannPreds))
   expect_true(length(subDannPreds) == nrow(xTest))
-  expect_true(all(colnames(subDannPreds) == "Class"))
 })
 
 subDannPreds <- sub_dann(xTrain, yTrain, xTest, 2, 50, 1, FALSE, FALSE, "mcd", 2)
@@ -318,7 +293,6 @@ test_that("Validate structure", {
   expect_true(is.vector(subDannPreds))
   expect_true(is.numeric(subDannPreds))
   expect_true(length(subDannPreds) == nrow(xTest))
-  expect_true(all(colnames(subDannPreds) == "Class"))
 })
 
 subDannPreds <- sub_dann(xTrain, yTrain, xTest, 2, 50, 1, FALSE, FALSE, "classical", 2)
@@ -326,7 +300,23 @@ test_that("Validate structure", {
   expect_true(is.vector(subDannPreds))
   expect_true(is.numeric(subDannPreds))
   expect_true(length(subDannPreds) == nrow(xTest))
-  expect_true(all(colnames(subDannPreds) == "Class"))
+})
+
+subDannPreds <- sub_dann(xTrain, yTrain, xTest, 2, 50, 1, FALSE, FALSE, "none", 2)
+test_that("Validate structure", {
+  expect_true(is.vector(subDannPreds))
+  expect_true(is.numeric(subDannPreds))
+  expect_true(length(subDannPreds) == nrow(xTest))
+})
+
+###############################################
+# default values match
+###############################################
+test_that("Defalut values match?", {
+  expect_true(formals(dann)$k == formals(sub_dann)$k)
+  expect_true(formals(dann)$neighborhood_size == formals(sub_dann)$neighborhood_size)
+  expect_true(formals(dann)$epsilon == formals(sub_dann)$epsilon)
+  expect_true(formals(dann)$probability == formals(sub_dann)$probability)
 })
 
 ###############################################